Fixing rtaudio-only references

+++ src/gui/vsDevice.cpp
@@ -227,7 +227,7 @@ void VsDevice::setServerId(QString serverId)
 
 // initJackTrip spawns a new jacktrip process with the desired settings
 JackTrip* VsDevice::initJackTrip(bool useRtAudio, std::string input, std::string output,
-                                 quint16 bufferSize, VsServerInfo* studioInfo)
+                                 int bufferSize, VsServerInfo* studioInfo)
 {
     m_jackTrip.reset(new JackTrip(JackTrip::CLIENTTOPINGSERVER, JackTrip::UDP, 2, 2,
 #ifdef WAIR  // wair
